WIP - Image Carousel.
Embedded and modal image carousels now properly receive imageUrl array.
