--- frontend/components/modal/media-image-carousel/ModalMediaImageCarousel.vue
@@ -1,1 +1,0 @@
-
